requirements.txt
# RA-FR Dependencies
# manually install:
# conda install faiss-gpu=1.7.2 mkl==2021.4.0 -c pytorch -c nvidia
# conda install pytorch==2.0.0 torchvision==0.15.1 torchaudio==2.0.1 pytorch-cuda=11.7 -c pytorch -c nvidia

# Core ML/DL libraries
numpy==1.26.3
scipy==1.13.1
pandas==2.0.2
scikit-learn==1.2.2
matplotlib==3.9.4
seaborn==0.13.2

# Deep learning & training
lightning==2.0.2
pytorch-lightning==2.3.3
timm==1.0.20
torchmetrics==1.8.2

# Experiment tracking & configuration
wandb==0.22.3
hydra-core==1.3.0

# Utilities
h5py==3.14.0
tqdm==4.65.0
easydict==1.13
matin==0.1
rich==12.3.0
PyYAML==6.0.3
pillow==11.0.0

# Development tools
ipdb==0.13.13
pretrainedmodels